requirements.txt
psycopg2-binary
redis
fastapi
uvicorn